requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
pillow>=11.3.0,<12.0.0
numpy>=2.2.6,<3.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0
homr>=0.6.0
